# docs/openapi.yaml
openapi: 3.0.2
servers:
  - url: https://api.mainnet.hiro.so/
    description: Mainnet
  - url: https://api.nakamoto.testnet.hiro.so/
    description: Nakamoto Testnet
  - url: https://api.testnet.hiro.so/
    description: Testnet
  - url: http://localhost:3999/
    description: Local
info:
  title: Stacks Blockchain API
  version: 'STACKS_API_VERSION'
  description: |
    Welcome to the API reference overview for the [Stacks Blockchain API](https://docs.hiro.so/stacks-blockchain-api).

    [Download Postman collection](https://hirosystems.github.io/stacks-blockchain-api/collection.json)
...
paths:
...
  /extended/v2/burn-blocks/{height_or_hash}/blocks:
    get:
      summary: Get blocks by burn block
      description: |
        Retrieves a list of blocks confirmed by a specific burn block
      tags:
        - Blocks
      operationId: get_blocks_by_burn_block
      parameters:
        - name: height_or_hash
          in: path
          description: filter by burn block height, hash, or the constant `latest` to filter for the most recent burn block
          required: true
          schema:
            oneOf:
              - type: integer
                example: 42000
              - type: string
                example: "0x4839a8b01cfb39ffcc0d07d3db31e848d5adf5279d529ed5062300b9f353ff79"
        - name: limit
          in: query
          description: max number of blocks to fetch
          required: false
          schema:
            type: integer
            example: 20
        - name: offset
          in: query
          description: index of first burn block to fetch
          required: false
          schema:
            type: integer
            example: 0
      responses:
        200:
          description: List of blocks
          content:
            application/json:
              schema:
                $ref: ./api/blocks/get-nakamoto-blocks.schema.json
              example:
                $ref: ./api/blocks/get-nakamoto-blocks.example.json
  /extended/v2/blocks:
    get:
      summary: Get blocks
      description: |
        Retrieves a list of recently mined blocks
      tags:
        - Blocks
      operationId: get_blocks
      parameters:
        - name: limit
          in: query
          description: max number of blocks to fetch
          required: false
          schema:
            type: integer
            example: 20
        - name: offset
          in: query
          description: index of first burn block to fetch
          required: false
          schema:
            type: integer
            example: 0
      responses:
        200:
          description: List of blocks
          content:
            application/json:
              schema:
                $ref: ./api/blocks/get-nakamoto-blocks.schema.json
              example:
                $ref: ./api/blocks/get-nakamoto-blocks.example.json
  /extended/v2/blocks/{height_or_hash}:
    get:
      summary: Get block
      description: |
        Retrieves a single block
      tags:
        - Blocks
      operationId: get_block
      parameters:
        - name: height_or_hash
          in: path
          description: filter by block height, hash, index block hash or the constant `latest` to filter for the most recent block
          required: true
          schema:
            oneOf:
              - type: integer
                example: 42000
              - type: string
                example: "0x4839a8b01cfb39ffcc0d07d3db31e848d5adf5279d529ed5062300b9f353ff79"
      responses:
        200:
          description: Block
          content:
            application/json:
              schema:
                $ref: ./entities/blocks/nakamoto-block.schema.json
              example:
                $ref: ./entities/blocks/nakamoto-block.example.json
  /extended/v2/blocks/{height_or_hash}/transactions:
    get:
      summary: Get transactions by block
      description: |
        Retrieves transactions confirmed in a single block
      tags:
        - Transactions
      operationId: get_transactions_by_block
      parameters:
        - name: height_or_hash
          in: path
          description: filter by block height, hash, index block hash or the constant `latest` to filter for the most recent block
          required: true
          schema:
            oneOf:
              - type: integer
                example: 42000
              - type: string
                example: "0x4839a8b01cfb39ffcc0d07d3db31e848d5adf5279d529ed5062300b9f353ff79"
      responses:
        200:
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: ./api/transaction/get-transactions.schema.json
              example:
                $ref: ./api/transaction/get-transactions.example.json
...
  /extended/v1/block:
    get:
      summary: Get recent blocks
      deprecated: true
      description: |
          **NOTE:** This endpoint is deprecated in favor of [Get blocks](/api/get-blocks).

          Retrieves a list of recently mined blocks

          If you need to actively monitor new blocks, we highly recommend subscribing to [WebSockets or Socket.io](https://github.com/hirosystems/stacks-blockchain-api/tree/master/client) for real-time updates.
      tags:
        - Blocks
      operationId: get_block_list
      parameters:
        - name: limit
          in: query
          description: max number of blocks to fetch
          required: false
          schema:
            type: integer
            default: 20
            maximum: 30
        - name: offset
          in: query
          description: index of first block to fetch
          required: false
          schema:
            type: integer
            example: 42000
      responses:
        200:
          description: List of blocks
          content:
            application/json:
              schema:
                $ref: ./api/blocks/get-blocks.schema.json
              example:
                $ref: ./api/blocks/get-blocks.example.json
  /extended/v1/block/{hash}:
    parameters:
      - name: hash
        in: path
        description: Hash of the block
        required: true
        schema:
          type: string
          example: "0x4839a8b01cfb39ffcc0d07d3db31e848d5adf5279d529ed5062300b9f353ff79"
    get:
      deprecated: true
      summary: Get block by hash
      description: |
        **NOTE:** This endpoint is deprecated in favor of [Get block](/api/get-block).
        
        Retrieves block details of a specific block for a given chain height. You can use the hash from your latest block ('get_block_list' API) to get your block details.
      tags:
        - Blocks
      operationId: get_block_by_hash
      responses:
        200:
          description: Block
          content:
            application/json:
              schema:
                $ref: ./entities/blocks/block.schema.json
              example:
                $ref: ./entities/blocks/block.example.json
        404:
          description: Cannot find block with given ID
          content:
            application/json:
              example:
                $ref: ./api/errors/block-not-found.example.json
...
  /rosetta/v1/block:
    post:
      tags:
        - Rosetta
      summary: Get a Block
      operationId: rosetta_block
      description: Retrieves the Block information for a given block identifier including a list of all transactions in the block.
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: ./api/rosetta/rosetta-block-response.schema.json
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: ./entities/rosetta/rosetta-error.schema.json
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./api/rosetta/rosetta-block-request.schema.json
            example: 
              $ref: ./api/rosetta/rosetta-block-request-body.example.json
...
  /extended/v1/tx/block/{block_hash}:
    get:
      deprecated: true
      operationId: get_transactions_by_block_hash
      summary: Transactions by block hash
      description: |
        **NOTE:** This endpoint is deprecated in favor of [Get transactions by block](/api/get-transactions-by-block).

        Retrieves a list of all transactions within a block for a given block hash.
      tags:
        - Transactions
      parameters:
        - name: block_hash
          in: path
          description: Hash of block
          required: true
          schema:
            type: string
            example: "0x0a83d82a65460a9e711f85a44616350280040b75317dbe486a923c1131b5ff99"
        - name: limit
          in: query
          description: max number of transactions to fetch
          required: false
          schema:
            type: integer
            example: 10
        - name: offset
          in: query
          description: index of first transaction to fetch
          required: false
          schema:
            type: integer
            example: 42000
      responses:
        200:
          description: List of Transactions
          content:
            application/json:
              schema:
                $ref: ./api/transaction/get-transactions.schema.json
              example:
                $ref: ./api/transaction/get-transactions.example.json
...